+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Update events for channel/empower spells were not behaving as intended
 - Hide changelog button on non UCB frames
 - Show/Hide preview was not working correctly in between castbar types of tabs
+- Copy profile was copying generated keys that were not in schema 
 
 ---